+++ tests/bases/test_collections.py
@@ -20,7 +20,17 @@ import torch
 
 from tests.helpers import seed_all
 from tests.helpers.testers import DummyMetricDiff, DummyMetricSum
-from torchmetrics import Accuracy, CohenKappa, ConfusionMatrix, F1Score, Metric, MetricCollection, Precision, Recall
+from torchmetrics import (
+    Accuracy,
+    CohenKappa,
+    ConfusionMatrix,
+    F1Score,
+    MatthewsCorrCoef,
+    Metric,
+    MetricCollection,
+    Precision,
+    Recall,
+)
 
 seed_all(42)
 
@@ -254,6 +264,8 @@ def test_collection_check_arg():
 
 
 def test_collection_filtering():
+    """Test that collections works with the kwargs argument."""
+
     class DummyMetric(Metric):
         def __init__(self):
             super().__init__()
@@ -310,11 +322,20 @@ def test_collection_filtering():
         ),
     ],
 )
-def test_check_compute_groups(metrics, expected):
+@pytest.mark.parametrize(
+    "prefix, postfix",
+    [
+        [None, None],
+        ["prefix_", None],
+        [None, "_postfix"],
+        ["prefix_", "_postfix"],
+    ],
+)
+def test_check_compute_groups(metrics, expected, prefix, postfix):
     """Check that compute groups are formed after initialization."""
-    m = MetricCollection(deepcopy(metrics), compute_groups=True)
+    m = MetricCollection(deepcopy(metrics), prefix=prefix, postfix=postfix, compute_groups=True)
     # Construct without for comparison
-    m2 = MetricCollection(deepcopy(metrics), compute_groups=False)
+    m2 = MetricCollection(deepcopy(metrics), prefix=prefix, postfix=postfix, compute_groups=False)
 
     assert len(m.compute_groups) == len(m)
     assert m2.compute_groups == {}
@@ -392,6 +413,24 @@ def test_compute_group_define_by_user():
     preds = torch.randn(10, 3).softmax(dim=-1)
     target = torch.randint(3, (10,))
     m.update(preds, target)
+    assert m.compute()
+
+
+def test_compute_on_different_dtype():
+    """Check that extraction of compute groups are robust towards difference in dtype."""
+    m = MetricCollection(
+        [
+            ConfusionMatrix(num_classes=3),
+            MatthewsCorrCoef(num_classes=3),
+        ]
+    )
+    assert not m._groups_checked
+    assert m.compute_groups == {0: ["ConfusionMatrix"], 1: ["MatthewsCorrCoef"]}
+    preds = torch.randn(10, 3).softmax(dim=-1)
+    target = torch.randint(3, (10,))
+    for _ in range(2):
+        m.update(preds, target)
+    assert m.compute_groups == {0: ["ConfusionMatrix", "MatthewsCorrCoef"]}
     assert m.compute()