+++ torchmetrics/audio/stoi.py
@@ -79,8 +79,9 @@ class ShortTimeObjectiveIntelligibility(Metric):
     """
     sum_stoi: Tensor
     total: Tensor
-    is_differentiable = False
-    higher_is_better = True
+    full_state_update: bool = False
+    is_differentiable: bool = False
+    higher_is_better: bool = True
 
     def __init__(
         self,