+++ torchmetrics/regression/cosine_similarity.py
@@ -51,8 +51,9 @@ class CosineSimilarity(Metric):
         tensor(0.8536)
 
     """
-    is_differentiable = True
-    higher_is_better = True
+    is_differentiable: bool = True
+    higher_is_better: bool = True
+    full_state_update: bool = False
     preds: List[Tensor]
     target: List[Tensor]