+++ torchmetrics/text/wil.py
@@ -50,8 +50,10 @@ class WordInfoLost(Metric):
         >>> metric(preds, target)
         tensor(0.6528)
     """
-    is_differentiable = False
-    higher_is_better = False
+    is_differentiable: bool = False
+    higher_is_better: bool = False
+    full_state_update: bool = False
+
     errors: Tensor
     target_total: Tensor
     preds_total: Tensor